+++ package/views/main_GUI/control_switches/task_switch.py
@@ -2,7 +2,6 @@ from threading import Thread
 from PyQt5 import QtGui
 from pycnbi import logger
 from random import randint
-import time
 
 class TaskSwitch():
     def onClicked_button_start_SV(self):
@@ -14,19 +13,6 @@ class TaskSwitch():
         self.is_experiment_on = True
         self.window.show()
 
-    def Time(self):
-        """
-        Call update subject view window image if task button clicked and display timer on experimenter
-        GUI.
-        """
-        os_time = time.time()
-        self.os_time_list.append(os_time)
-        if self.is_experiment_on:
-            self.Update_SV_image()
-        # Variables.add_one_run_time_counter()
-        # time_show = Variables.get_run_time_counter()
-        self.time_show += 1
-        self.ui.lcdNumber_timer.display(self.time_show)
 
     def Update_SV_image(self):
         """ Update subject view GUI visual cues inlcuding Idle, focus, prepare, two, one, task"""
